merge

+++ oop-bound/base_block.cpp
@@ -2,12 +2,13 @@
 
 BaseBlock::BaseBlock()
 {
+
 	position.x = position.y = 0;
 	if (!texture.loadFromFile("image/base.png")) {
 		//error
 	}
 	sprite.setTexture(texture, true);
-	sprite.setPosition(BLOCK_SIZE * position.x, BLOCK_SIZE * position.y);
+	sprite.setPosition((float)BLOCK_SIZE * position.x, (float)BLOCK_SIZE * position.y);
 }
 
 BaseBlock::~BaseBlock()
@@ -29,5 +30,5 @@ void BaseBlock::setPosition(int x, int y)
 {
 	position.x = x;
 	position.y = y;
-	sprite.setPosition(BLOCK_SIZE * x, BLOCK_SIZE * y);
+	sprite.setPosition((float)BLOCK_SIZE * x, (float)BLOCK_SIZE * y);
 }
